+++ routes/api.php
@@ -1,8 +1,15 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use TomatoPHP\FilamentCmsApi\Http\Controllers\CategoryController;
+use TomatoPHP\FilamentCmsApi\Http\Controllers\PostController;
 
-Route::middleware(['auth:sanctum'])->prefix('api/pages')->name('api.pages.')->group(function () {
-    Route::get('/', [\TomatoPHP\FilamentCms\Http\Controllers\PageController::class, 'index'])->name('index');
-    Route::get('/{model}', [\TomatoPHP\FilamentCms\Http\Controllers\PageController::class, 'show'])->name('show');
-});
+Route::middleware(config('filament-cms-api.middleware', ['api']))
+    ->prefix(config('filament-cms-api.prefix', 'api/cms'))
+    ->name('filament-cms-api.')
+    ->group(function () {
+        Route::get('posts', [PostController::class, 'index'])->name('posts.index');
+        Route::get('posts/{slug}', [PostController::class, 'show'])->name('posts.show');
+        Route::get('categories', [CategoryController::class, 'index'])->name('categories.index');
+        Route::get('categories/{slug}', [CategoryController::class, 'show'])->name('categories.show');
+    });